Beginning to overcome quite a nasty bug: Ivy hides irrelevant model facts when representing the model as clauses, which is nice but introduces an abstraction which is problematic when we decompose a cex to a concrete *trace*. Consider the program: action coolest_check_renaming = {    local i1: iterator, i2: iterator {        assume ~cme(I);        assume i1 ~= i2;        call check_renaming(i2, i1);        #cme(I) := false;        call check_renaming(i1, i2);        if (~cme(I)) {            assume true        } else {            assume true        }    } } export coolest_check_renaming
action check_renaming(i1: iterator, i3: iterator) = {
  if (i1 = i3) {
    cme(i1) := true
  }
}

which is erroneously reported as unsafe, but when with the commented line uncommented (cme(I) := false) it is considered the safe.
The reason is that the decomposition returns two transition (one for each call) but doesn't report on which of these cme(I) became true, because cme seems irrelevant to Ivy because it doesn't appear anywhere in the clauses. None of these transitions can be blocked, because if cme(I) doesn't change then the transition is valid but can be blocked if cme(I) does become true.

To circumvent this we concretize the transition by assuming that relations that don't appear anywhere are false, and hope that this concretization will not hurt our generalization where it's not essential due to the use of unsat_core.

In this commit adding information just on the cme() relation, to be completed for all relation symbols.